requirements.txt
absl-py==2.1.0
anndata==0.11.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
array_api_compat==1.9.1
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
async-lru==2.0.4
attrs==24.2.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
beautifulsoup4==4.12.3
biopython==1.83
bleach==6.1.0
blis==1.2.0
build==1.2.2.post1
cachetools==5.5.0
catalogue==2.0.10
catboost==1.2.7
certifi==2024.2.2
cffi==1.17.1
charset-normalizer==3.3.2
chroma-hnswlib==0.7.6
chromadb==0.5.20
click==8.1.7
cloudpathlib==0.20.0
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
contourpy==1.2.1
cycler==0.12.1
cymem==2.0.11
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.15
distlib==0.3.8
Django==5.0.4
dnspython==2.7.0
docutils==0.21.2
durationpy==0.9
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et-xmlfile==1.1.0
executing==2.1.0
fastapi==0.115.5
fastjsonschema==2.20.0
filelock==3.16.1
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
fsspec==2024.10.0
gast==0.6.0
gdown==5.2.0
google-ai-generativelanguage==0.6.15
google-api-core==2.24.0
google-api-python-client==2.159.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.4
google-pasta==0.2.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.69.0
grpcio-status==1.69.0
h11==0.14.0
h5py==3.11.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.26.2
humanfriendly==10.0
idna==3.8
imageio==2.35.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
imutils==0.5.4
ipykernel==6.29.5
ipython==8.27.0
ipywidgets==8.1.5
isoduration==20.11.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jedi==0.19.1
Jinja2==3.1.4
joblib==1.4.2
json5==0.9.25
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.14.2
jupyter_server_terminals==0.5.3
jupyterlab==4.2.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keras==3.6.0
keyring==25.4.1
kiwisolver==1.4.5
kubernetes==31.0.0
langcodes==3.5.0
language_data==1.3.0
lazy_loader==0.4
legacy-api-wrap==1.4
libclang==18.1.1
lightgbm==4.5.0
llvmlite==0.43.0
lxml==5.3.0
marisa-trie==1.2.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplot==0.1.9
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.1
mmh3==5.0.1
modeller==0.1.1
monotonic==1.6
more-itertools==10.5.0
mpmath==1.3.0
murmurhash==1.0.12
mysql-connector-python==8.3.0
mysqlclient==2.2.4
namex==0.0.8
natsort==8.4.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
neo4j==5.26.0
nest-asyncio==1.6.0
networkx==3.3
nh3==0.2.18
notebook==7.2.2
notebook_shim==0.2.4
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.20.1
opencv-python==4.10.0.84
openpyxl==3.1.5
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-instrumentation-asgi==0.49b2
opentelemetry-instrumentation-fastapi==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
opentelemetry-util-http==0.49b2
opt-einsum==3.3.0
optree==0.12.1
orjson==3.10.12
overrides==7.7.0
packaging==24.1
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.4
patsy==1.0.1
pdfkit==1.0.0
pexpect==4.9.0
pickle-mixin==1.0.2
pillow==10.3.0
pkginfo==1.10.0
platformdirs==4.3.6
plotly==5.24.1
posthog==3.7.2
preshed==3.0.9
prettytable==3.10.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
proto-plus==1.25.0
protobuf==5.28.3
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.1
pydantic_core==2.27.1
Pygments==2.18.0
pyloco==0.0.139
pymongo==4.10.1
pynndescent==0.5.13
pyparsing==3.1.2
PyPDF2==3.0.1
PyPika==0.48.9
pyproject_hooks==1.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==2.0.7
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
readme_renderer==44.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==2.0.0
rfc3986-validator==0.1.1
rich==13.8.1
rpds-py==0.20.0
rsa==4.9
safetensors==0.4.5
scanpy==1.10.4
scikit-image==0.24.0
scikit-learn==1.5.1
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.3.1
session_info==1.0.0
setuptools==75.1.0
shapely==2.0.6
shellingham==1.5.4
SimpleWebSocketServer==0.1.2
six==1.16.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.5.1
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.4.0
terminado==0.18.1
tf_keras==2.18.0
thinc==8.3.4
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.3.0
tokenizers==0.20.3
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.5
traitlets==5.14.3
transformers==4.46.3
twine==5.1.1
typer==0.13.1
types-python-dateutil==2.9.0.20241003
typing==3.7.4.3
typing_extensions==4.12.2
tzdata==2024.1
umap-learn==0.5.6
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.2
ushlex==0.99.1
uvicorn==0.32.1
uvloop==0.21.0
virtualenv==20.26.5
wasabi==1.1.3
watchfiles==1.0.0
wcwidth==0.2.13
weasel==0.4.1
webcolors==24.8.0
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
Werkzeug==3.0.4
wheel==0.44.0
widgetsnbextension==4.0.13
wrapt==1.16.0
xgboost==2.1.1
xlrd==2.0.1
zipp==3.20.2
